split contributor audio section into different accordians
