+++ website/src/components/edit-word-audio/index.tsx
@@ -1,4 +1,5 @@
 import { ReactElement } from "react"
+import { WordAudio } from "src/panel-layout"
 import { UserRole, useCognitoUserGroups, useUserRole } from "../../auth"
 import * as Dailp from "../../graphql/dailp"
 import { ContributorEditWordAudio } from "./contributor"
@@ -7,17 +8,17 @@ import { EditorEditWordAudio } from "./editor"
 export const EditWordAudio = (p: {
   word: Dailp.FormFieldsFragment
 }): ReactElement => {
-  const groups = useCognitoUserGroups()
-  const userRole = useUserRole()
+  const role = useUserRole()
 
-  console.log("Branching based on", { groups, userRole })
+  console.log("Branching based on", { role })
 
-  switch (userRole) {
-    case UserRole.CONTRIBUTOR:
-      return <ContributorEditWordAudio word={p.word} />
-    case UserRole.EDITOR:
+  switch (role) {
+    case UserRole.Editor:
       return <EditorEditWordAudio word={p.word} />
-    case UserRole.READER:
-      return <div>You do not have permission to edit word audio.</div>
+
+    case UserRole.Contributor:
+      return <ContributorEditWordAudio word={p.word} />
+    case UserRole.Reader:
+      return <WordAudio word={p.word} />
   }
 }
